Move readFirstBytes method as requested
I fully agree that the previous function order was not logical. But
I found more logical to have "public" (i.e. "exported") variables
first (because they are what the developers are looking for at first
glance), and then put below the "private" and last "tree-leaf" functions
and variables. The longer one reads into the file, the deeper in the
dependency chain variables one finds.