Fix error handling
So far promise could be resolved with any value apart of Error instance
and rejected only with Error instance.
This fix allows to resolve with any value including error and reject
with any value.
While it's definitely not recommended to reject with non error values,
it matches `throw` behavior (any type of value can be thrown)